RFCT: refactor model bar into subfeatures format, extract api into own file and update hooks
