+++ frontend/src/features/model-bar/subfeatures/utility-panel/hooks/useModelFormValidate.js
@@ -1,6 +1,6 @@
-import { z } from "zod";
-import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { z } from "zod";
 
 const useModelFormValidate = () => {
   const modelFormInfo = {
